# Hexadecimal optimizations