--- a/newsfragments/223.performance.rst
+++ b/newsfragments/223.performance.rst
@@ -1,0 +1,1 @@
+Performance improvement of up to 65% on ``is_0x_prefixed``